conf-service/src/main/java/com/endava/internship/cryptomarket/confservice/business/validators/ValidUsernameValidator.java
package com.endava.internship.cryptomarket.confservice.business.validators;

import com.endava.internship.cryptomarket.confservice.business.validators.annotations.ValidUsername;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;

import static java.util.regex.Pattern.matches;

public class ValidUsernameValidator implements ConstraintValidator<ValidUsername, String> {

    @Override
    public boolean isValid(final String username, final ConstraintValidatorContext constraintValidatorContext) {
        return matches("^(?=.{4,32}$)(?![_])(?!.*[_]{2})[A-Za-z0-9_]+(?<![_])$", username);
    }
}
